Commands: Separate each Command into classes (contd.) Commands are all handled in a singular execute method, with a long switch/case staement block handlung all commands case-by-case.
The execute() method has multiple code smells, such as weak SLAP and being "too long" of a method.

The previous commit was pushed prematurely, thus the entire implementation was not pushed. This commit contains the missing changes.
Let's continue to fix code quality by making the Command class an abstract class, and seperating the handling of each Command type into its own class, named as <commandName>Command. The execute() method is turned into an abstract method, so that each type of Command can have their own implementation of execute().
