+++ src/main/java/MayoBot/commands/UnknownCommand.java
@@ -1,4 +1,16 @@
-package MayoBot.exceptions;
+package MayoBot.commands;
 
-public class UnknownCommand {
+import MayoBot.exceptions.MayoBotException;
+import MayoBot.exceptions.UnknownCommandException;
+import MayoBot.task.TaskList;
+import MayoBot.ui.Ui;
+
+public class UnknownCommand extends Command {
+    public UnknownCommand(String command, String arguments) {
+        super(command, arguments);
+    }
+
+    public String execute(Ui ui, TaskList taskList, boolean isGui) throws MayoBotException {
+        throw new UnknownCommandException(this.getCommand());
+    }
 }
